Separando action Types
