+++ src/reducers/cycles.ts
@@ -1,0 +1,56 @@
+export interface Cycle {
+  id: string
+  task: string
+  minutAmount: number
+  date: Date
+  interrupDate?: Date
+  finishedDate?: Date
+}
+
+interface CycleState {
+  cycles: Cycle[]
+  activeCycleId: string | null
+}
+
+export enum ActionTypes {
+  CREATED_NEW_CYCLE_CONTEXT = 'CREATED_NEW_CYCLE_CONTEXT',
+  INTERRUPTED_CYCLES_CONTEXT = 'INTERRUPTED_CYCLES_CONTEXT',
+  MARK_CURRENT_CYCLE_AS_FINISHED = 'MARK_CURRENT_CYCLE_AS_FINISHED',
+}
+
+export function CyclesReducers(state: CycleState, action: any) {
+  switch (action.type) {
+    case ActionTypes.CREATED_NEW_CYCLE_CONTEXT:
+      return {
+        ...state,
+        cycles: [...state.cycles, action.payload.newCycle],
+        activeCycleId: action.payload.newCycle.id,
+      }
+    case ActionTypes.INTERRUPTED_CYCLES_CONTEXT:
+      return {
+        ...state,
+        cycles: state.cycles.map((cycle) => {
+          if (cycle.id === state.activeCycleId) {
+            return { ...cycle, interrupDate: new Date() }
+          } else {
+            return cycle
+          }
+        }),
+        activeCycleId: null,
+      }
+    case ActionTypes.MARK_CURRENT_CYCLE_AS_FINISHED:
+      return {
+        ...state,
+        cycles: state.cycles.map((cycle) => {
+          if (cycle.id === state.activeCycleId) {
+            return { ...cycle, finishedDate: new Date() }
+          } else {
+            return cycle
+          }
+        }),
+        activeCycleId: null,
+      }
+    default:
+      return state
+  }
+}
